Sample from the target distribution using MCMC

Sample from the target distribution using MCMC

run_mcmc(
  data,
  sample_ids,
  loci,
  is_missing = FALSE,
  thin = 1,
  burnin = 10000,
  samples = 10000,
  verbose = TRUE,
  allele_freq_threshold = 1e-05,
  eps_pos_0 = 0.005,
  eps_pos_var = 1,
  eps_pos_alpha = 0.5,
  eps_pos_beta = 99.5,
  eps_neg_0 = 0.005,
  eps_neg_var = 1,
  eps_neg_alpha = 0.5,
  eps_neg_beta = 99.5,
  max_eps_pos = 2,
  max_eps_neg = 2,
  max_coi = 20,
  allele_freq_vars = 1,
  adapt_allele_freq_vars = TRUE
)

Arguments

data

Data to be used in MCMC, expected input is a list of lists of numeric vectors, where each list element is a collection of observations across samples at a single genetic locus.

sample_ids

Vector containing the ordered unique identifier for samples

loci

Vector containing the ordered unique identifer for loci

is_missing

Boolean matrix indicating whether the observation should be treated as missing data and ignored. Number of rows equals the number of loci, number of columns equals the number samples. Alternatively, the user may pass in FALSE if no data should be considered missing.

thin

Positive Integer. How often to sample from mcmc, 1 means do not thin

burnin

Positive Integer. Number of MCMC samples to discard as burnin

samples

Positive Integer. Number of samples to take after burnin

verbose

Logical indicating if progress is printed

allele_freq_threshold

0-1 Numeric. Lowest allowed value for an allele frequency.

eps_pos_0

Numeric. Initial eps_pos value

eps_pos_var

Numeric. Variance used in sampling eps_pos

eps_neg_0

Numeric. Initial eps_neg value

eps_neg_var

Numeric. Variance used in sampling eps_neg

max_eps_pos

Numeric. Maximum allowed value for eps_pos

max_eps_neg

Numeric. Maximum allowed value for eps_neg

max_coi

Positive Numeric. Maximum allowed complexity of infection

allele_freq_vars

Positive Numeric. Variance used in sampling allele frequencies

adapt_allele_freq_vars

Logical indicating whether to adapt variance to achieve an acceptance rate of .27

eps_pos_shape

Positive Numeric. Shape parameter in Gamma distribution for eps_pos prior

eps_pos_scale

Positive Numeric. Scale parameter in Gamma distribution for eps_pos prior

eps_neg_shape

Positive Numeric. Shape parameter in Gamma distribution for eps_neg prior

eps_neg_scale

Positive Numeric. Scale parameter in Gamma distribution for eps_neg prior