Sample from the target distribution using MCMC

Sample from the target distribution using MCMC

run_mcmc(
  data,
  sample_ids,
  loci,
  is_missing = FALSE,
  thin = 1,
  burnin = 10000,
  samples = 10000,
  complexity_limit = 2050,
  importance_sampling_depth = 300,
  importance_sampling_scaling_factor = 100,
  verbose = TRUE,
  eps_pos_0 = 0.01,
  eps_pos_var = 0.001,
  eps_pos_alpha = 1,
  eps_pos_beta = 99,
  eps_neg_0 = 0.05,
  eps_neg_var = 0.005,
  eps_neg_alpha = 5,
  eps_neg_beta = 95,
  max_eps_pos = 0.5,
  max_eps_neg = 0.5,
  mean_coi_prior_shape = 1.5,
  mean_coi_prior_scale = 0.5,
  mean_coi_var = 1,
  allele_freq_var = 0.1
)

Arguments

data
Data to be used in MCMC, expected input is a list of lists of numeric vectors, where each list element is a collection of observations across samples at a single genetic locus.

sample_ids
Vector containing the ordered unique identifier for samples

loci
Vector containing the ordered unique identifer for loci

is_missing
Boolean matrix indicating whether the observation should be treated as missing data and ignored. Number of rows equals the number of loci, number of columns equals the number samples. Alternatively, the user may pass in FALSE if no data should be considered missing.

thin
Positive Integer. How often to sample from mcmc, 1 means do not thin

burnin
Positive Integer. Number of MCMC samples to discard as burnin

samples
Positive Integer. Number of samples to take after burnin

complexity_limit
Limit on the total number of computations before resorting to an importance sampling based approach of the integral. Total number of computations is approximately

importance_sampling_depth
Positive Integer. Min number of samples to take during importance sampling. Larger values result in longer computation time, too small of values will result in poor mixing.

importance_sampling_scaling_factor
Positive Ingeter. Total additional importance samples taken per increase in complexity of infection. Total number of samples = importance_sampling_depth + coi * importance_sampling_scaling_ factor

verbose
Logical indicating if progress is printed

eps_pos_0
0-1 Numeric. Initial eps_pos value

eps_pos_var
0-1 Numeric. Variance used in sampling eps_pos

eps_pos_alpha
Positive Numeric. Alpha parameter in Beta distribution for eps_pos prior

eps_pos_beta
Positive Numeric. Beta parameter in Beta distribution for eps_pos prior

eps_neg_0
0-1 Numeric. Initial eps_neg value

eps_neg_var
0-1 Numeric. Variance used in sampling eps_neg

eps_neg_alpha
Positive Numeric. Alpha parameter in Beta distribution for eps_neg prior

eps_neg_beta
Positive Numeric. Beta parameter in Beta distribution for eps_neg prior

max_eps_pos
0-1 Numeric. Maximum allowed value for eps_pos

max_eps_neg
0-1 Numeric. Maximum allowed value for eps_neg

mean_coi_prior_shape
Positive Numeric. Shape parameter for gamma prior on mean complexity of infection.

mean_coi_prior_scale
Positive Numeric. Scale parameter for gamma prior on mean complexity of infection.

mean_coi_var
Positive Numeric. Variance used in sampling mean_coi

allele_freq_var
Positive Numeric. Variance used in sampling allele frequencies