Simulate data generated according to the assumed model

Simulate data generated according to the assumed model

simulate_data(
  mean_coi,
  num_samples,
  epsilon_pos,
  epsilon_neg,
  locus_freq_alphas = NULL,
  allele_freqs = NULL,
  internal_relatedness = 0
)

Arguments

mean_coi

Mean multiplicity of infection drawn from a Poisson

num_samples

Total number of biological samples to simulate

epsilon_pos

False positive rate, expected number of false positives

epsilon_neg

False negative rate, expected number of false negatives

locus_freq_alphas

List of alpha vectors to be used to simulate from a Dirichlet distribution to generate allele frequencies.

allele_freqs

List of allele frequencies to be used instead of simulating allele frequencies

Value

Simulated data that is structured to go into the MCMC sampler